+++ package.json
@@ -27,6 +27,7 @@
     "jose": "^5.2.3",
     "jsonwebtoken": "^9.0.2",
     "jspdf": "^2.5.1",
+    "jwt-decode": "^4.0.0",
     "lucide-react": "^0.367.0",
     "next": "^14.2.0",
     "primeflex": "^3.3.1",
