+++ src/app/(users)/mycourses/page.js
