+++ src/app/admin/dashboard/page.js
@@ -1,0 +1,8 @@
+
+export default function dashboardPage() {
+  return (
+    <div>
+      <h1>Hello dashboardPage</h1>
+    </div>
+  );
+}
