+++ src/app/courses/explore/layout.js
@@ -1,7 +1,7 @@
 'use client'
 
-import AsideStudent from "@/components/AsideStudent/AsideStudent";
-import { NavCourses } from "@/components/NavCourses/NavCourses";
+import AsideStudent from "@/components/usersComponents/AsideStudent/AsideStudent";
+import { NavCourses } from "@/components/usersComponents/NavCourses/NavCourses";
 import styles from './LayoutCourses.module.scss'
 
 export default function CoursesLayout({ children }) {
