+++ src/app/page.js
@@ -1,9 +1,9 @@
 'use client'
-import CardCategories from '@/components/CardCategories/CardCategories'
-import { Footer } from '@/components/Footer/Footer'
-import { NavHome } from '@/components/Nav/NavHome'
-import SignUpCards from '@/components/SignUpCard/SignUpCards'
-import TopMessageHome from '@/components/TopMessageHome/TopMessageHome'
+import CardCategories from '@/components/usersComponents/CardCategories/CardCategories'
+import { Footer } from '@/components/usersComponents/Footer/Footer'
+import { NavHome } from '@/components/usersComponents/Nav/NavHome'
+import SignUpCards from '@/components/usersComponents/SignUpCard/SignUpCards'
+import TopMessageHome from '@/components/usersComponents/TopMessageHome/TopMessageHome'
 import styles from '@/styles/HomePage.module.scss'
 import Image from 'next/image'
 import { useEffect, useState } from 'react'
