+++ src/components/usersComponents/AsideStudent/AsideStudent.js
@@ -12,7 +12,7 @@ function AsideStudent() {
         <div className={styles.aside}>
             <ul>
                 <li><a href="/courses/explore" className={`items-center justify-start ${pathname.includes('/courses/explore') ? styles.active : ''}`}><Compass /> <p>Explorar</p></a></li>
-                <li><a href="/admin/profile#ultimoCurso" className={`items-center justify-start ${pathname.includes('/courses/mycourses') ? styles.active : ''}`}><BookCheck /> <p>Mis cursos</p></a></li>
+                <li><a href="/profile#ultimoCurso" className={`items-center justify-start ${pathname.includes('/courses/mycourses') ? styles.active : ''}`}><BookCheck /> <p>Mis cursos</p></a></li>
                 <li><a href="/courses/myprogress" className={`items-center justify-start ${pathname.includes('/courses/stats') ? styles.active : ''}`}><GraduationCap /> <p>Mi progreso</p></a></li>
             </ul>
