+++ src/components/usersComponents/Nav/NavHome.js
@@ -3,7 +3,7 @@ import React from 'react'
 import Image from 'next/image'
 import { useRouter } from 'next/navigation'
 import { useSelector, useDispatch } from 'react-redux'
-import { login, logout } from '../../features/auth/loginSlice'
+import { login, logout } from '../../../features/auth/loginSlice'
 import { getCookie, deleteCookie } from 'cookies-next';
 import { Bell, ChevronDown, Mail } from 'lucide-react';
 import { Navbar, NavbarBrand, NavbarContent, NavbarItem, Link, DropdownItem, DropdownTrigger, Dropdown, DropdownMenu, Avatar } from "@nextui-org/react";
@@ -64,7 +64,7 @@ export const NavHome = () => {
               </div>
             </DropdownTrigger>
             <DropdownMenu aria-label="Profile Actions" variant="flat">
-              <DropdownItem key="settings" href='/admin/profile'>Configuración</DropdownItem>
+              <DropdownItem key="settings" href='/profile'>Configuración</DropdownItem>
               <DropdownItem key="logout" color="danger">
                 <div onClick={handleLogout}>
                   Cerrar sesión
